Project Management - Subway Security System
Lead a project of the team of 6 members during our project management program at Humber College and made a Project Charted in the supervision of our prof. Nadine Blakes.

The main purpose of the project is to create a secure and safe environment for both commuters and operators, by undertaking the necessary measures in terms of infrastructure, technology, and manpower. Incorporating the optimal use of available plans, procedures, resources, and methodology place a key role in achieving our goal.